fix a error in token expired

+++ pkg/jwt/jwt.go
@@ -47,7 +47,7 @@ func ParseToken(tokenStr string) (string, int, error) {
 
 	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
 		err = nil
-		if claims["exp"].(int64) < time.Now().Unix() {
+		if int64(claims["exp"].(float64)) < time.Now().Unix() {
 			err = &TokenExpiredError{}
 		}
